Fix a couple of potential issues in SFTP rmtree()
This commit changes the STP rmtree() method to limit the number of
concurrent directory listing requests to avoid potentially hitting a
limit on the number of open file handles that SFTP servers allow in a
single session.

Also, exception handling has been improved to avoid potentially seeing
"Future exception was never retrieved" errors when multiple exceptions
are raised during its parallel execution. Only the first exception will
be reported back to the caller, but the others will be retrieved and
discarded. If a caller wants to see all the exceptions, they can do so
by specifying a callback in the "onerror" argument.

+++ asyncssh/sftp.py
@@ -2661,23 +2661,29 @@ class SFTPClient:
             tasks = []
 
             try:
-                async for entry in self.scandir(path):
-                    if entry.filename in (b'.', b'..'):
-                        continue
+                async with sem:
+                    async for entry in self.scandir(path):
+                        if entry.filename in (b'.', b'..'):
+                            continue
 
-                    mode = entry.attrs.permissions
-                    entry = posixpath.join(path, entry.filename)
+                        mode = entry.attrs.permissions
+                        entry = posixpath.join(path, entry.filename)
 
-                    if stat.S_ISDIR(mode):
-                        task = _rmtree(entry)
-                    else:
-                        task = _unlink(entry)
+                        if stat.S_ISDIR(mode):
+                            task = _rmtree(entry)
+                        else:
+                            task = _unlink(entry)
 
-                    tasks.append(asyncio.ensure_future(task))
+                        tasks.append(asyncio.ensure_future(task))
             except SFTPError:
                 onerror(self.scandir, path, sys.exc_info())
 
-            await asyncio.gather(*tasks)
+            results = await asyncio.gather(*tasks, return_exceptions=True)
+            exc = next((result for result in results
+                        if isinstance(result, Exception)), None)
+
+            if exc:
+                raise exc
 
             try:
                 await self.rmdir(path)
@@ -2694,6 +2700,7 @@ class SFTPClient:
         # pylint: enable=function-redefined
 
         path = self.encode(path)
+        sem = asyncio.Semaphore(_MAX_SFTP_REQUESTS)
 
         try:
             if await self.islink(path):
